added berserk added bloodlust added hammer of light added earthliving + windfury added holy radiance aded inspiring presence added last stand added light infusion added lightning rod added some specs added cooldowns changed from double to float

+++ src/com/ebicep/warlords/util/Utils.java
@@ -10,7 +10,7 @@ public class Utils {
         Location eye = player.getEyeLocation();
         eye.setY(eye.getY() + .5);
         Vector toEntity = player1.getEyeLocation().toVector().subtract(eye.toVector());
-        double dot = toEntity.normalize().dot(eye.getDirection());
+        float dot = (float) toEntity.normalize().dot(eye.getDirection());
 
         return dot > 0.98D;
     }
